+++ lib/async_redux.dart
@@ -1,7 +1,7 @@
 export 'src/event_redux.dart';
+export 'src/local_persist.dart';
 export 'src/navigate_action.dart';
 export 'src/persistor.dart';
-export 'src/saver_loader.dart';
 export 'src/store.dart';
 export 'src/store_tester.dart';
 export 'src/user_exception.dart';